requirement.txt
tqdm==4.62.3
requests==2.24.0
pycryptodome==3.10.4